move catalogue.js
move to templates folder

move catalogue.js
